Added functions to handle part of the use's ENV
The 'build_path' function creates a linked list of the user's PATH
which is used by the parser to ensure the line receieved on the shell
is an actual command and exists in the PATH before sending it off to
the command executor. Also the '_getenv' mimics the original getenv()
function.