Added the parser for the shell
This file contains multiple functions that processes the command
provided by the user to ensure it is valid before doing anything
else. It takes care of tokenizing it and performing checks before
sending them to the executor or the relevant functions. It also has
enough error checking and handling implemented.