Updated from _free_on_exit() to multi_free()
This update modified the  function '_free_on_exit()' to 'multi_free()'.
Previously, I thought that was all I was going to use it for, but
I've learned better. I created it to free a variable number of
dynamically allocated objects on 'exit', but it's proving useful for
general purpose memory deallocations too. I now use it to free multiple
variables in one line as opposed to the one line for each in the past.